requirement.txt
requests
google-generativeai
pyTelegramBotAPI
python-dotenv

